Change mappings for comma
From <bslash> (\) to single quote (')
Comma is used by defaul to search within line (for a letter) in oposite
direction, some times it is very useful.
And <bslash> I use as something like 'localleader' key. In this
connection when you press <bslash> there is a small delay that is
unwanted if you just want to quickly go to a particular place.